+++ cmd/jaeger/internal/integration/e2e_integration.go
@@ -24,7 +24,6 @@ import (
 	"github.com/jaegertracing/jaeger/cmd/jaeger/internal/integration/storagecleaner"
 	"github.com/jaegertracing/jaeger/plugin/storage/integration"
 	"github.com/jaegertracing/jaeger/ports"
-	"github.com/jaegertracing/jaeger/storage_v2/v1adapter"
 )
 
 const otlpPort = 4317
@@ -150,10 +149,9 @@ func (s *E2EStorageIntegration) e2eInitialize(t *testing.T, storage string) {
 
 	s.SpanWriter, err = createSpanWriter(logger, otlpPort)
 	require.NoError(t, err)
-	// TODO upgrade to APIv3 client
-	spanReader, err := createSpanReader(logger, ports.QueryGRPC)
+	traceReader, err := createTraceReader(logger, ports.QueryGRPC)
 	require.NoError(t, err)
-	s.TraceReader = v1adapter.NewTraceReader(spanReader)
+	s.TraceReader = traceReader
 
 	t.Cleanup(func() {
 		// Call e2eCleanUp to close the SpanReader and SpanWriter gRPC connection.
@@ -210,9 +208,7 @@ func (s *E2EStorageIntegration) doHealthCheck(t *testing.T) bool {
 // e2eCleanUp closes the SpanReader and SpanWriter gRPC connection.
 // This function should be called after all the tests are finished.
 func (s *E2EStorageIntegration) e2eCleanUp(t *testing.T) {
-	spanReader, err := v1adapter.GetV1Reader(s.TraceReader)
-	require.NoError(t, err)
-	require.NoError(t, spanReader.(io.Closer).Close())
+	require.NoError(t, s.TraceReader.(io.Closer).Close())
 	require.NoError(t, s.SpanWriter.(io.Closer).Close())
 }
