--- cmd/jaeger/internal/integration/grpc_trace_reader.go
@@ -1,169 +1,0 @@
-// Copyright (c) 2024 The Jaeger Authors.
-// SPDX-License-Identifier: Apache-2.0
-
-package integration
-
-import (
-	"context"
-	"errors"
-	"fmt"
-	"io"
-	"iter"
-	"math"
-
-	"go.opentelemetry.io/collector/pdata/ptrace"
-	"go.uber.org/zap"
-	"google.golang.org/grpc"
-	"google.golang.org/grpc/credentials/insecure"
-
-	"github.com/jaegertracing/jaeger/internal/proto/api_v3"
-	"github.com/jaegertracing/jaeger/model"
-	"github.com/jaegertracing/jaeger/pkg/iter"
-	"github.com/jaegertracing/jaeger/ports"
-	"github.com/jaegertracing/jaeger/proto-gen/api_v2"
-	"github.com/jaegertracing/jaeger/storage/spanstore"
-	"github.com/jaegertracing/jaeger/storage_v2/tracestore"
-)
-
-var (
-	_ tracestore.Reader = (*traceReader)(nil)
-	_ io.Closer         = (*traceReader)(nil)
-)
-
-// traceReader retrieve span data from Jaeger-v2 query with api_v2.QueryServiceClient.
-type traceReader struct {
-	logger     *zap.Logger
-	clientConn *grpc.ClientConn
-	client     api_v3.QueryServiceClient
-}
-
-func createTraceReader(logger *zap.Logger, port int) (*traceReader, error) {
-	logger.Info("Creating the span reader", zap.Int("port", port))
-	opts := []grpc.DialOption{
-		grpc.WithTransportCredentials(insecure.NewCredentials()),
-	}
-
-	cc, err := grpc.NewClient(ports.PortToHostPort(port), opts...)
-	if err != nil {
-		return nil, err
-	}
-
-	return &traceReader{
-		logger:     logger,
-		clientConn: cc,
-		client:     api_v3.NewQueryServiceClient(cc),
-	}, nil
-}
-
-func (r *traceReader) Close() error {
-	r.logger.Info("Closing the gRPC trace reader")
-	return r.clientConn.Close()
-}
-
-func (r *traceReader) GetTraces(ctx context.Context, traceIDs ...tracestore.GetTraceParams) iter.Seq2[[]ptrace.Traces, error] {
-	// model.TraceIDFromOTEL()
-	// TODO api_v3 does not have multi-get-trace
-	stream, err := r.client.GetTrace(ctx, &api_v3.GetTraceRequest{
-		TraceID:   query.TraceID,
-		StartTime: query.StartTime,
-		EndTime:   query.EndTime,
-	})
-	if err != nil {
-		return nil, unwrapNotFoundErr(err)
-	}
-
-	var spans []*model.Span
-	for received, err := stream.Recv(); !errors.Is(err, io.EOF); received, err = stream.Recv() {
-		if err != nil {
-			return nil, unwrapNotFoundErr(err)
-		}
-		for i := range received.Spans {
-			spans = append(spans, &received.Spans[i])
-		}
-		// r.logger.Info(fmt.Sprintf("GetTrace received %d spans (total %d)", len(received.Spans), len(spans)))
-	}
-	r.logger.Info(fmt.Sprintf("GetTraces received a total of %d spans", len(spans)))
-
-	return &model.Trace{
-		Spans: spans,
-	}, nil
-}
-
-func (r *traceReader) GetServices(ctx context.Context) ([]string, error) {
-	res, err := r.client.GetServices(ctx, &api_v2.GetServicesRequest{})
-	if err != nil {
-		return []string{}, err
-	}
-	return res.Services, nil
-}
-
-func (r *traceReader) GetOperations(ctx context.Context, query spanstore.OperationQueryParameters) ([]spanstore.Operation, error) {
-	var operations []spanstore.Operation
-	res, err := r.client.GetOperations(ctx, &api_v2.GetOperationsRequest{
-		Service:  query.ServiceName,
-		SpanKind: query.SpanKind,
-	})
-	if err != nil {
-		return operations, err
-	}
-	for _, operation := range res.Operations {
-		operations = append(operations, spanstore.Operation{
-			Name:     operation.Name,
-			SpanKind: operation.SpanKind,
-		})
-	}
-	return operations, nil
-}
-
-func (r *traceReader) FindTraces(ctx context.Context, query *spanstore.TraceQueryParameters) ([]*model.Trace, error) {
-	var traces []*model.Trace
-
-	if query.NumTraces > math.MaxInt32 {
-		return traces, fmt.Errorf("NumTraces must not greater than %d", math.MaxInt32)
-	}
-	stream, err := r.client.FindTraces(ctx, &api_v2.FindTracesRequest{
-		Query: &api_v2.TraceQueryParameters{
-			ServiceName:   query.ServiceName,
-			OperationName: query.OperationName,
-			Tags:          query.Tags,
-			StartTimeMin:  query.StartTimeMin,
-			StartTimeMax:  query.StartTimeMax,
-			DurationMin:   query.DurationMin,
-			DurationMax:   query.DurationMax,
-			//nolint: gosec // G115
-			SearchDepth: int32(query.NumTraces),
-		},
-	})
-	if err != nil {
-		return traces, err
-	}
-
-	totalSpans := 0
-	spanMaps := map[string][]*model.Span{}
-	for received, err := stream.Recv(); !errors.Is(err, io.EOF); received, err = stream.Recv() {
-		if err != nil {
-			return nil, unwrapNotFoundErr(err)
-		}
-		for i, span := range received.Spans {
-			traceID := span.TraceID.String()
-			if _, ok := spanMaps[traceID]; !ok {
-				spanMaps[traceID] = make([]*model.Span, 0)
-			}
-			spanMaps[traceID] = append(spanMaps[traceID], &received.Spans[i])
-		}
-		totalSpans += len(received.Spans)
-		r.logger.Info(fmt.Sprintf("FindTraces received %d spans (total %d)", len(received.Spans), totalSpans))
-	}
-	r.logger.Info(fmt.Sprintf("FindTraces received a total of %d spans", totalSpans))
-
-	for _, spans := range spanMaps {
-		traces = append(traces, &model.Trace{
-			Spans: spans,
-		})
-	}
-	return traces, nil
-}
-
-func (*traceReader) FindTraceIDs(context.Context, *spanstore.TraceQueryParameters) ([]model.TraceID, error) {
-	panic("not implemented")
-}
