+++ cmd/jaeger/internal/integration/trace_reader.go
@@ -1,0 +1,168 @@
+// Copyright (c) 2024 The Jaeger Authors.
+// SPDX-License-Identifier: Apache-2.0
+
+package integration
+
+import (
+	"context"
+	"errors"
+	"fmt"
+	"io"
+	"math"
+
+	"go.opentelemetry.io/collector/pdata/pcommon"
+	"go.opentelemetry.io/collector/pdata/ptrace"
+	"go.uber.org/zap"
+	"google.golang.org/grpc"
+	"google.golang.org/grpc/credentials/insecure"
+
+	"github.com/jaegertracing/jaeger/internal/proto/api_v3"
+	"github.com/jaegertracing/jaeger/model"
+	"github.com/jaegertracing/jaeger/pkg/iter"
+	"github.com/jaegertracing/jaeger/ports"
+	"github.com/jaegertracing/jaeger/storage/spanstore"
+	"github.com/jaegertracing/jaeger/storage_v2/tracestore"
+)
+
+var (
+	_ tracestore.Reader = (*traceReader)(nil)
+	_ io.Closer         = (*traceReader)(nil)
+)
+
+// traceReader retrieve span data from Jaeger-v2 query with api_v2.QueryServiceClient.
+type traceReader struct {
+	logger     *zap.Logger
+	clientConn *grpc.ClientConn
+	client     api_v3.QueryServiceClient
+}
+
+func createTraceReader(logger *zap.Logger, port int) (*traceReader, error) {
+	logger.Info("Creating the span reader", zap.Int("port", port))
+	opts := []grpc.DialOption{
+		grpc.WithTransportCredentials(insecure.NewCredentials()),
+	}
+
+	cc, err := grpc.NewClient(ports.PortToHostPort(port), opts...)
+	if err != nil {
+		return nil, err
+	}
+
+	return &traceReader{
+		logger:     logger,
+		clientConn: cc,
+		client:     api_v3.NewQueryServiceClient(cc),
+	}, nil
+}
+
+func (r *traceReader) Close() error {
+	r.logger.Info("Closing the gRPC trace reader")
+	return r.clientConn.Close()
+}
+
+func (r *traceReader) GetTraces(
+	ctx context.Context,
+	traceIDs ...tracestore.GetTraceParams,
+) iter.Seq2[[]ptrace.Traces, error] {
+	return func(yield func([]ptrace.Traces, error) bool) {
+		// api_v3 does not support multi-get, so loop through IDs
+		for _, idParams := range traceIDs {
+			stream, err := r.client.GetTrace(ctx, &api_v3.GetTraceRequest{
+				TraceId:   model.TraceIDFromOTEL(idParams.TraceID).String(),
+				StartTime: idParams.Start,
+				EndTime:   idParams.End,
+			})
+			if err != nil {
+				err = unwrapNotFoundErr(err)
+				if errors.Is(err, spanstore.ErrTraceNotFound) {
+					continue
+				}
+				yield(nil, err)
+				return
+			}
+			for received, err := stream.Recv(); !errors.Is(err, io.EOF); received, err = stream.Recv() {
+				if err != nil {
+					err = unwrapNotFoundErr(err)
+					if errors.Is(err, spanstore.ErrTraceNotFound) {
+						continue
+					}
+					yield(nil, err)
+					return
+				}
+				traces := received.ToTraces()
+				if !yield([]ptrace.Traces{traces}, nil) {
+					return
+				}
+			}
+		}
+	}
+}
+
+func (r *traceReader) GetServices(ctx context.Context) ([]string, error) {
+	res, err := r.client.GetServices(ctx, &api_v3.GetServicesRequest{})
+	if err != nil {
+		return []string{}, err
+	}
+	return res.Services, nil
+}
+
+func (r *traceReader) GetOperations(ctx context.Context, query tracestore.OperationQueryParameters) ([]tracestore.Operation, error) {
+	var operations []tracestore.Operation
+	res, err := r.client.GetOperations(ctx, &api_v3.GetOperationsRequest{
+		Service:  query.ServiceName,
+		SpanKind: query.SpanKind,
+	})
+	if err != nil {
+		return operations, err
+	}
+	for _, operation := range res.Operations {
+		operations = append(operations, tracestore.Operation{
+			Name:     operation.Name,
+			SpanKind: operation.SpanKind,
+		})
+	}
+	return operations, nil
+}
+
+func (r *traceReader) FindTraces(
+	ctx context.Context,
+	query tracestore.TraceQueryParams,
+) iter.Seq2[[]ptrace.Traces, error] {
+	return func(yield func([]ptrace.Traces, error) bool) {
+		if query.NumTraces > math.MaxInt32 {
+			yield(nil, fmt.Errorf("NumTraces must not be greater than %d", math.MaxInt32))
+			return
+		}
+		stream, err := r.client.FindTraces(ctx, &api_v3.FindTracesRequest{
+			Query: &api_v3.TraceQueryParameters{
+				ServiceName:   query.ServiceName,
+				OperationName: query.OperationName,
+				Attributes:    query.Tags,
+				StartTimeMin:  query.StartTimeMin,
+				StartTimeMax:  query.StartTimeMax,
+				DurationMin:   query.DurationMin,
+				DurationMax:   query.DurationMax,
+				//nolint: gosec // G115
+				SearchDepth: int32(query.NumTraces),
+				RawTraces:   true,
+			},
+		})
+		if err != nil {
+			yield(nil, unwrapNotFoundErr(err))
+			return
+		}
+		for received, err := stream.Recv(); !errors.Is(err, io.EOF); received, err = stream.Recv() {
+			if err != nil {
+				yield(nil, unwrapNotFoundErr(err))
+				return
+			}
+			traces := received.ToTraces()
+			if !yield([]ptrace.Traces{traces}, nil) {
+				return
+			}
+		}
+	}
+}
+
+func (*traceReader) FindTraceIDs(ctx context.Context, query tracestore.TraceQueryParams) iter.Seq2[[]pcommon.TraceID, error] {
+	panic("not implemented")
+}
